Removed unused variable $tax_total in function invoice_item_tax() in sql_queries.php
In include/sql_queries.php the function invoice_item_tax() refers to a variable $tax_total. It does not seem to be computed or initialised anywhere. Only subsequent computation as an accumulator is seen.

Even if it were initialised, where is it used?
It does not figure in any INSERT statement or as an argument to an function in there.