Simplify extensions with report & menu items plus generic PDO requests.
These changes allow multiple extensions for new reports or new items
that will appear in the menu line to be developed. Before this change,
the entire template/default/reports/index.tpl file for reports and the
entire templates/default/menu.tpl file for menus had to be included in
the extension. This update modifies the logice so that only the portion
of the index or menu for the extension is specified and through use of
section identifiers, that portion is merged to the index or menu when
they are built. This means you can have multiple extensions for reports
and/or menu items that do not impact each other as they did prior to
this change.

Also, a function to dynamically build the PDO <i>INSERT</i>,
<i>UPDATE</i> and <i>DELETE</i> statement from the item values in the
<i>$_POST</i> super global is included. This will allo extensions (or
any code) that maintains a database table to perform the maintenance
request without the need to add functions to the <i>sql_queries.php</i>
file.