Shift popup calendar to the right so it doesn't hide it icon.
If you use a start and end date that are positioned vertically, the
popup for the start date covers the icon to access the popup for the
stop date making it a two click process to get to the stop date when no
change is made in the start date. This change positions the popup to the
right so the icons aren't obscured and the stop date can be accessed
directly.