Changes to support extension development without overriding key files.
These changes impliment a process that allows files that might be
updated by multiple extentions to specify only the section that needs to
be changed rather than having to replicate the entire file. This means
there won't be competition between between extensions having to consider
whether or not another extension is enabled.