Add check number field to payments.
This extension was updated to add support for a check_number field in
payments. If the extension is enabled, the first access of payment
information will add the check_number field to the si_payment table.
Additionally, if check type payments have a numeric value in the first
characters of the notes field, they will be moved to the check_number
field. Also, the payment receipt template has been updated to report
biller information adjacent to the customer information, enhancing the
appearence of the printed output.