Add support for "company_name" to languages files.
This change is in preparation for implimentation of change to allow
users to specify the name of their company for display on screens in
place of the "Simple Invoices" text the currently displays. This
includes breaking the "simple_invoices_powered_by" language item into
two parts, "powered_by" and the already existing "simple_invoices".