User Security enhancements
The primary user security change is to add username as a unique user id
in place of the email address currently used. The password security is
upgraded with the use of minimum length and required character types
optionally configured for each SI instance.

View the user_security extension README file for more details on all
enhancements.