Move sql_queries.php functions to respective classes.
A lot of changes rolled into this. Primary is moving of functions from
sql_queries.php into respective classes as static functions. This is an
intermediate step to upgrade PHP version and support libraries such as
Zend and Smarty.

This update also fixes the problem of not being able to download a PDF,
DOC or XLS file.

Note that not all configuration options have been tested, so their could
be problems with certain functionality.