Add holiday logo feature.
This feature allows you to have special logos for invoices and payment
receipts for different holidays. Currently New Years (January),
Valentines (February), Easter (April), Independance Day (July),
Thanksgiving (November) and Christmas (December). The logic takes the
logo file name defined in the Biller's record and if the current date is
in one of the holiday months, inserts the holiday on the file name and
checks to see if it exists. Values inserted are "_newyears",
"_valentines", "_easter", "_independence", "_thanksgiving" and
"_christmas". If the file exists, the holiday logo is used. To use this,
modify the {$logo|urlsafe} line in your template to be {holiday_logo
logo=$logo|urlsafe}. EX: my_logo.png becomes my_logo_christmas.png for
December. If the modified file does not exist, the default file is used.