Added message logging level control to tmp/log/si.log output.
Added debug logging feature that can be turned on and off through
setting of new value in the config/config.php file. If you use
custom.config.php (as recommended), copy the "zend.logger_level" line
and its comments from the config/config.php file from this update. The
default level is EMERG. The most detailed level is DEBUG which should be
used if you are having issues and no error is generated.