Changes to support PHP 7.0x.
Functions supporing magic quotes were deprecated in PHP 5.4x and removed
in PHP 7.0x. All logic referencing these functions are now preceeded by
a test to see if the function exists. If not, it is not performed. If it
is, it is performed. The bottom line is that this code now works in both
PHP 7 and previous versions.